requirements.txt
asgiref==3.8.1
coverage==7.6.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==71.1.0
sqlparse==0.5.1
uritemplate==4.1.1
